+++ clients/socks5/src/commands/init.rs
@@ -14,8 +14,9 @@ use nym_bin_common::output_format::OutputFormat;
 use nym_client_core::client::base_client::storage::gateway_details::OnDiskGatewayDetails;
 use nym_client_core::client::key_manager::persistence::OnDiskKeys;
 use nym_client_core::config::GatewayEndpointConfig;
+use nym_client_core::error::ClientCoreError;
 use nym_client_core::init::helpers::current_gateways;
-use nym_client_core::init::GatewaySetup;
+use nym_client_core::init::types::{GatewayDetails, GatewaySetup};
 use nym_crypto::asymmetric::identity;
 use nym_sphinx::addressing::clients::Recipient;
 use nym_topology::NymTopology;
@@ -113,7 +114,7 @@ impl From<Init> for OverrideConfig {
 #[derive(Debug, Serialize)]
 pub struct InitResults {
     #[serde(flatten)]
-    client_core: nym_client_core::init::InitResults,
+    client_core: nym_client_core::init::types::InitResults,
     socks5_listening_port: u16,
     client_address: String,
 }
@@ -121,7 +122,7 @@ pub struct InitResults {
 impl InitResults {
     fn new(config: &Config, address: &Recipient, gateway: &GatewayEndpointConfig) -> Self {
         Self {
-            client_core: nym_client_core::init::InitResults::new(
+            client_core: nym_client_core::init::types::InitResults::new(
                 &config.core.base,
                 address,
                 gateway,
@@ -213,8 +214,7 @@ pub(crate) async fn execute(args: Init) -> Result<(), Socks5ClientError> {
         Some(&network_gateways),
     )
     .await
-    .tap_err(|err| eprintln!("Failed to setup gateway\nError: {err}"))?
-    .details;
+    .tap_err(|err| eprintln!("Failed to setup gateway\nError: {err}"))?;
 
     // TODO: ask the service provider we specified for its interface version and set it in the config
 
@@ -229,7 +229,10 @@ pub(crate) async fn execute(args: Init) -> Result<(), Socks5ClientError> {
 
     let address = init_details.client_address()?;
 
-    let init_results = InitResults::new(&config, &address, &init_details.gateway_details);
+    let GatewayDetails::Configured(gateway_details) = init_details.gateway_details else {
+        return Err(ClientCoreError::UnexpectedPersistedCustomGatewayDetails)?
+    };
+    let init_results = InitResults::new(&config, &address, &gateway_details);
     println!("{}", args.output.format(&init_results));
 
     Ok(())
