+++ common/client-core/src/error.rs
@@ -123,6 +123,9 @@ pub enum ClientCoreError {
     #[error("unable to upgrade config file to `{new_version}`")]
     UnableToUpgradeConfigFile { new_version: String },
 
+    #[error("the provided gateway details don't much the stored data")]
+    MismatchedStoredGatewayDetails,
+
     #[error("custom selection of gateway was expected")]
     CustomGatewaySelectionExpected,
